Split code into multiple smaller files and fix the imports in the top-module __init__ file to accomodate this. Close #38
