+++ addon/i3dio/node_classes/__init__.py
@@ -1,0 +1,5 @@
+from . import node
+from . import shape
+from . import merge_group
+from . import file
+from . import material
