+++ addon/i3dio/properties.py
