+++ addon/i3dio/ui_export.py
@@ -13,7 +13,7 @@ from bpy.types import (
     Panel
 )
 
-from . import exporter, xml_i3d, i3d_properties
+from . import exporter, xml_i3d, properties
 
 
 @orientation_helper(axis_forward='-Z', axis_up='Y')
